Prepare campaign for C23_WMDE_iPad_01
Ticket: https://phabricator.wikimedia.org/T349700
