--- banners/pad/styles/BannerVarHeadline.scss
@@ -1,36 +1,0 @@
-@use 'src/themes/Treedip/variables/colors';
-@use 'src/themes/Treedip/variables/fonts';
-
-.wmde-banner {
-	&-content {
-		&-headline {
-			font-family: fonts.$content;
-			font-size: 21px;
-			font-weight: bold;
-			text-align: center;
-			margin-top: 20px;
-			padding: 0 52px;
-			position: relative;
-
-			&::before {
-				content: '';
-				position: absolute;
-				height: 1px;
-				top: 50%;
-				left: 52px;
-				right: 52px;
-				background: colors.$primary;
-				z-index: 98;
-			}
-
-			&-text {
-				position: relative;
-				display: inline-block;
-				z-index: 99;
-				padding: 0 10px;
-				line-height: 24px;
-				background: colors.$white;
-			}
-		}
-	}
-}
