--- banners/pad/styles/styles_var.scss
@@ -1,40 +1,0 @@
-// This is the file where we import the theme-specific component styles
-@use 'src/components/BannerConductor/banner-transition';
-@use './Banner';
-@use 'MainBanner' with (
-	$banner-height: 353px,
-	$form-width: 256px
-);
-@use './BannerVarHeadline';
-@use 'src/themes/UseOfFunds/UseOfFunds';
-@use 'src/themes/Treedip/defaults';
-@use 'src/themes/Treedip/ButtonClose/ButtonClose';
-@use 'src/themes/Treedip/ProgressBar/ProgressBar' with (
-	$progress-bar-margin: 0 15px
-);
-@use 'src/themes/Treedip/DonationForm/DonationForm' with (
-	$button-font-size: 12px
-);
-@use 'src/themes/Treedip/DonationForm/MultiStepDonation';
-@use 'src/themes/Treedip/DonationForm/SubComponents/SelectGroup';
-@use 'src/themes/Treedip/DonationForm/SubComponents/SelectGroupButtons' with (
-	$height: 30px
-);
-@use 'src/themes/Treedip/DonationForm/SubComponents/SelectCustomAmountButton' with (
-	$height: 30px
-);
-@use 'src/themes/Treedip/DonationForm/SubComponents/SmsBox';
-@use 'src/themes/Treedip/DonationForm/Forms/MainDonationForm' with (
-	$padding: 40px 0 0
-);
-@use 'src/themes/Treedip/DonationForm/Forms/UpgradeToYearlyButtonForm';
-@use 'src/themes/Treedip/DonationForm/Forms/CustomAmountForm';
-@use 'src/themes/Treedip/Footer/FooterAlreadyDonated';
-@use 'src/themes/Treedip/Footer/BannerFooter' with (
-	$right-column-width: 256px
-);
-@use 'src/themes/Treedip/Footer/SelectionInput';
-@use 'src/themes/Treedip/Message/Message';
-@use 'src/themes/Treedip/Slider/KeenSlider';
-@use 'src/themes/Treedip/AlreadyDonatedModal/AlreadyDonatedModal';
-@use 'src/themes/Treedip/SoftClose/SoftClose';
