+++ campaign_info.toml
@@ -60,9 +60,9 @@ orientation = [ "portrait", "landscape"]
 [pad]
 name = "iPad"
 icon = "pad"
-campaign = "C23_WMDE_iPad_00"
-description = "VAR has a header text above the slider"
-campaign_tracking = "pad00-ba-230426"
+campaign = "C23_WMDE_iPad_01"
+description = "VAR has payment logos"
+campaign_tracking = "pad01-ba-231030"
 preview_link = "/wiki/Wikipedia:Hauptseite?devbanner={{banner}}&banner=B22_WMDE_local_prototype"
 preview_url = 'https://de.wikipedia.org/wiki/Wikipedia:Hauptseite?banner={{banner}}&devMode'
 wrapper_template = "wikipedia_org"
@@ -70,13 +70,13 @@ use_of_funds_source = "MediaWiki:WMDE_Fundraising/UseOfFunds_2023_DE"
 
 [pad.banners.ctrl]
 filename = "./banners/pad/banner_ctrl.ts"
-pagename = "B23_WMDE_iPad_00_ctrl"
-tracking = "org-pad00-230501-ctrl"
+pagename = "B23_WMDE_iPad_01_ctrl"
+tracking = "org-pad01-231030-ctrl"
 
 [pad.banners.var]
 filename = "./banners/pad/banner_var.ts"
-pagename = "B23_WMDE_iPad_00_var"
-tracking = "org-pad00-230501-var"
+pagename = "B23_WMDE_iPad_01_var"
+tracking = "org-pad01-231030-var"
 
 [pad.test_matrix]
 device = [ 'ipad_mini', 'ipad', 'ipad_pro_9_7_inch', 'ipad_pro_12_inch' ]
