--- test/banners/pad/components/BannerVar.spec.ts
@@ -1,129 +1,0 @@
-import { afterEach, beforeEach, describe, test } from 'vitest';
-import { mount, VueWrapper } from '@vue/test-utils';
-import Banner from '../../../../banners/pad/components/BannerVar.vue';
-import { BannerStates } from '@src/components/BannerConductor/StateMachine/BannerStates';
-import { newDynamicContent } from '@test/banners/dynamicCampaignContent';
-import { useOfFundsContent } from '@test/banners/useOfFundsContent';
-import { formItems } from '@test/banners/formItems';
-import { CurrencyEn } from '@src/utils/DynamicContent/formatters/CurrencyEn';
-import { TrackerStub } from '@test/fixtures/TrackerStub';
-import { softCloseFeatures } from '@test/features/SoftCloseDesktop';
-import { useOfFundsFeatures } from '@test/features/UseOfFunds';
-import { bannerContentAnimatedTextFeatures, bannerContentFeatures } from '@test/features/BannerContent';
-import { alreadyDonatedModalFeatures } from '@test/features/AlreadyDonatedModal';
-import { resetFormModel } from '@test/resetFormModel';
-import { useFormModel } from '@src/components/composables/useFormModel';
-import { donationFormFeatures } from '@test/features/forms/MainDonation_UpgradeToYearlyButton';
-import { DynamicContent } from '@src/utils/DynamicContent/DynamicContent';
-import { bannerMainFeatures } from '@test/features/MainBanner';
-
-const formModel = useFormModel();
-const translator = ( key: string ): string => key;
-
-describe( 'BannerVar.vue', () => {
-	let wrapper: VueWrapper<any>;
-	beforeEach( () => {
-		resetFormModel( formModel );
-	} );
-
-	const getWrapper = ( dynamicContent: DynamicContent = null ): VueWrapper<any> => {
-		// attachTo the document body to fix an issue with Vue Test Utils where
-		// clicking a submit button in a form does not fire the submit event
-		wrapper = mount( Banner, {
-			attachTo: document.body,
-			props: {
-				bannerState: BannerStates.Pending,
-				useOfFundsContent
-			},
-			global: {
-				mocks: {
-					$translate: translator
-				},
-				provide: {
-					translator: { translate: translator },
-					dynamicCampaignText: dynamicContent ?? newDynamicContent(),
-					formActions: { donateWithAddressAction: 'https://example.com', donateWithoutAddressAction: 'https://example.com' },
-					currencyFormatter: new CurrencyEn(),
-					formItems,
-					tracker: new TrackerStub()
-				}
-			}
-		} );
-
-		return wrapper;
-	};
-
-	afterEach( () => {
-		wrapper.unmount();
-	} );
-
-	describe( 'Main Banner', () => {
-		test.each( [
-			[ 'expectDoesNotEmitCloseEvent' ]
-		] )( '%s', async ( testName: string ) => {
-			await bannerMainFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Content', () => {
-		test.each( [
-			[ 'expectSlideShowPlaysWhenBecomesVisible' ],
-			[ 'expectSlideShowStopsOnFormInteraction' ]
-		] )( '%s', async ( testName: string ) => {
-			await bannerContentFeatures[ testName ]( getWrapper() );
-		} );
-
-		test.each( [
-			[ 'expectHidesAnimatedVisitorsVsDonorsSentenceInSlideShow' ],
-			[ 'expectShowsAnimatedVisitorsVsDonorsSentenceInSlideShow' ]
-		] )( '%s', async ( testName: string ) => {
-			await bannerContentAnimatedTextFeatures[ testName ]( getWrapper );
-		} );
-	} );
-
-	describe( 'Donation Form Happy Paths', () => {
-		test.each( [
-			[ 'expectMainDonationFormSubmitsWhenSofortIsSelected' ],
-			[ 'expectMainDonationFormSubmitsWhenYearlyIsSelected' ],
-			[ 'expectMainDonationFormGoesToUpgrade' ],
-			[ 'expectUpgradeToYearlyFormSubmitsUpgrade' ],
-			[ 'expectUpgradeToYearlyFormSubmitsDontUpgrade' ],
-			[ 'expectUpgradeToYearlyFormGoesToMainDonation' ]
-		] )( '%s', async ( testName: string ) => {
-			await donationFormFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Soft Close', () => {
-		test.each( [
-			[ 'expectShowsSoftClose' ],
-			[ 'expectEmitsSoftCloseCloseEvent' ],
-			[ 'expectEmitsSoftCloseMaybeLaterEvent' ],
-			[ 'expectEmitsSoftCloseTimeOutEvent' ],
-			[ 'expectEmitsBannerContentChangedOnSoftClose' ]
-		] )( '%s', async ( testName: string ) => {
-			await softCloseFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Use of Funds', () => {
-		test.each( [
-			[ 'expectShowsUseOfFunds' ],
-			[ 'expectHidesUseOfFunds' ]
-		] )( '%s', async ( testName: string ) => {
-			await useOfFundsFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-	describe( 'Already Donated Modal', () => {
-		test.each( [
-			[ 'expectShowsAlreadyDonatedModal' ],
-			[ 'expectHidesAlreadyDonatedModal' ],
-			[ 'expectFiresMaybeLaterEvent' ],
-			[ 'expectFiresGoAwayEvent' ]
-		] )( '%s', async ( testName: string ) => {
-			await alreadyDonatedModalFeatures[ testName ]( getWrapper() );
-		} );
-	} );
-
-} );
